Adds dockerfile and updates OpenCV dependency
This adds a dockerfile, called `Dockerfile` in the
root folder, that builds an image containing the
compiled OceanCV library. All of the `WITH_*`
flags can be set in the file to control the build
targets. Some minor code changes were made to be
compatible with a newer version of OpenCV.